+++ test/unit/canvas.js
@@ -392,7 +392,7 @@
     assert.equal(activeSelection[1], rect1, 'then rect1');
   });
 
-  QUnit.test('start multiselection: preserve', function (assert) {
+  QUnit.test('start multiselection: canvas-stacking', function (assert) {
     var rect1 = new fabric.Rect();
     var rect2 = new fabric.Rect();
     canvas.add(rect1, rect2);
@@ -434,7 +434,6 @@
     updateActiveSelection(canvas, [rect1, rect2], rect3, 'selection-order');
     assert.equal(isFired, true, 'selected on rect3 fired');
   });
-
 
   QUnit.test('continuing multiselection respects order of objects', function (assert) {
     const rect1 = new fabric.Rect();
@@ -517,6 +516,19 @@
     canvas.add(rect1, rect2, rect3);
     updateActiveSelection(canvas, [rect1, rect2, rect3], canvas.getActiveSelection(), 'selection-order');
     assert.deepEqual(canvas.getActiveObjects(), [rect1, rect2, rect3], 'nothing happened');
+    assert.ok(canvas.getActiveSelection() === canvas.getActiveObject(), 'still selected');
+  });
+
+  QUnit.test('multiselection: selecting a target behind active selection', assert => {
+    const rect1 = new fabric.Rect({ left: 10, width: 10, height: 10 });
+    const rect2 = new fabric.Rect({ width: 10, height: 10 });
+    const rect3 = new fabric.Rect({ top: 10, width: 10, height: 10 });
+    canvas.add(rect1, rect2, rect3);
+    initActiveSelection(canvas, rect1, rect3);
+    assert.ok(canvas.getActiveSelection() === canvas.getActiveObject(), 'selected');
+    assert.deepEqual(canvas.getActiveObjects(), [rect1, rect3], 'created');
+    canvas.__onMouseDown({ clientX: 7, clientY: 7, [canvas.selectionKey]: true });
+    assert.deepEqual(canvas.getActiveObjects(), [rect1, rect2, rect3], 'added from behind active selection');
     assert.ok(canvas.getActiveSelection() === canvas.getActiveObject(), 'still selected');
   });