+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [next]
 
 - fix(): interact with a selected `perPixelTargetFind` object [#9167](https://github.com/fabricjs/fabric.js/pull/9167)
+- CD(): website submodule [#9165](https://github.com/fabricjs/fabric.js/pull/9165)
+
+## [6.0.0-beta12]
+
+- fix(Object): border rendering with padding under group [#9161](https://github.com/fabricjs/fabric.js/pull/9161)
+- fix(MultiSelection): add target from behind active selection [#8744](https://github.com/fabricjs/fabric.js/issues/8744)
+- test(): fix snapshots by removing version [#9164](https://github.com/fabricjs/fabric.js/pull/9164)
 
 ## [6.0.0-beta11]